backend/requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
numpy
python-dotenv
google-generativeai
